requirements.txt
pdfplumber>=0.10.0
pymupdf>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
docxtpl>=0.16.0
python-docx>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
